Declare php extension Fileinfo as requirement
It's used in `src/Bex/Behat/ScreenshotExtension/Driver/Local.php`. If your
php-version doesn't have the extension installed, you'll run into an error
otherwise. Because the error requires some effort to be tracked down, it
is now declared as a requirement in composer.json.

That way, composer will report the missing extension.